test_storage_web now runnable on Python 3 (even if it doesn't pass).
